Kool20

See merge request khuloodbatis/twitter-app!16

+++ app/Http/Controllers/Tweets/TweetLikeController.php
@@ -11,22 +11,22 @@ use App\Http\Resources\Tweet\TweetResource;
 
 class TweetLikeController extends Controller
 {
-    public function store(Tweet $tweet, Request $request)
+    public function store(Request $request, Tweet $tweet)
     {
-
-        $tweet->likes()->where('user_id',Auth::id())->exists();
-
-        $tweet->likes()->create([
+        $isLiked = $tweet->likes()->where('user_id', Auth::id())->exists();
+        if (!$isLiked) {
+            $tweet->likes()->create([
                 'user_id' => $request->user()->id,
             ]);
+        }
+
         return  new TweetResource($tweet);
     }
 
-    public function destroy(Tweet $tweet, Request $request)
+    public function destroy(Request $request, Tweet $tweet)
     {
-        $tweet->likes()->delete([
-            'user_id' => $request->user()->id,
-        ]);
+        $tweet->likes()->where('user_id', $request->user()->id)->delete();
 
+        return response()->json(['status' => 'like was deleted']);
     }
 }
